requirements.txt
dht11
RPi.GPIO
pandas
paho-mqtt
